[techDebt]  melhorias e inclusões de rotas para maior controle de crud de usuário , adicionado testes unitários para novos métodos e melhorias no tratamento de exceções

+++ Postech.NETT11.PhaseOne/Postech.NETT11.PhaseOne.WebApp/Endpoints/GameRoute.cs
@@ -13,6 +13,11 @@ public class GameRoute:BaseRoute
     {
         Route = "/game";
     }
+
+    private bool IsAdmin(HttpContext context)
+    {
+        return context.User.IsInRole("Admin");
+    }
     protected override void AddRoute(RouteGroupBuilder group)
     {
         group.MapGet("/", GetAllGames)
@@ -41,14 +46,20 @@ public class GameRoute:BaseRoute
             .RequireAuthorization("Admin");
     }
 
-    private async Task<NoContent> DeleteGameAsync(Guid id, IGameService service)
+    private async Task<NoContent> DeleteGameAsync(Guid id, HttpContext context, IGameService service)
     {
+        if (!IsAdmin(context))
+            throw new UnauthorizedAccessException();
+        
         var deleteResult = await service.DeleteGameAsync(id);
         return TypedResults.NoContent();
     }
 
-    private async Task<Ok<GameResponse>> UpdateGameAsync(Guid id,UpdateGameRequest request, IGameService service)
+    private async Task<Ok<GameResponse>> UpdateGameAsync(Guid id, HttpContext context, UpdateGameRequest request, IGameService service)
     {
+        if (!IsAdmin(context))
+            throw new UnauthorizedAccessException();
+        
         var updatedGame = await service.UpdateGameAsync(id,request);
         return TypedResults.Ok(updatedGame);
     }
@@ -70,8 +81,11 @@ public class GameRoute:BaseRoute
         return TypedResults.Ok(games.ToList());
     }
 
-    private async Task<Ok<GameResponse>> CreateGameAsync(CreateGameRequest request, IGameService service)
+    private async Task<Ok<GameResponse>> CreateGameAsync(HttpContext context, CreateGameRequest request, IGameService service)
     {
+        if (!IsAdmin(context))
+            throw new UnauthorizedAccessException();
+        
         var newGame = await service.AddGameAsync(request);
         return TypedResults.Ok(newGame);
     }
